requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-dotenv
google-generativeai
scaledown==0.1.4